Move auth-template files into subdirectory and extract shared HTTP capture helper
