+++ packages/core/src/auth-template/get-auth-template.js
@@ -1,24 +1,21 @@
 'use strict';
 
-const applyMiddleware = require('./middleware');
-const ensureArray = require('./tools/ensure-array');
+const applyMiddleware = require('../middleware');
+const ensureArray = require('../tools/ensure-array');
 
 // before middlewares
-const addBasicAuthHeader = require('./http-middlewares/before/add-basic-auth-header');
-const addQueryParams = require('./http-middlewares/before/add-query-params');
-const createInjectInputMiddleware = require('./http-middlewares/before/inject-input');
-const prepareRequest = require('./http-middlewares/before/prepare-request');
-const sanitizeHeaders = require('./http-middlewares/before/sanatize-headers');
+const addBasicAuthHeader = require('../http-middlewares/before/add-basic-auth-header');
+const addQueryParams = require('../http-middlewares/before/add-query-params');
+const createInjectInputMiddleware = require('../http-middlewares/before/inject-input');
+const prepareRequest = require('../http-middlewares/before/prepare-request');
+const sanitizeHeaders = require('../http-middlewares/before/sanatize-headers');
 
-const { REPLACE_CURLIES } = require('./constants');
+const { REPLACE_CURLIES } = require('../constants');
+const { withHttpCapture } = require('./http-capture');
 
-const http = require('http');
-const https = require('https');
 const vm = require('vm');
 const lodash = require('lodash');
-const { Readable } = require('stream');
-const { EventEmitter } = require('events');
-const errors = require('./errors');
+const errors = require('../errors');
 
 // --- Helpers ---
 
@@ -678,87 +675,6 @@ const runTestFunctionSurvival = async (
     };
   };
 
-  const origHttpRequest = http.request;
-  const origHttpsRequest = https.request;
-  const origHttpGet = http.get;
-  const origHttpsGet = https.get;
-
-  const patchedRequest = (origFn, protocol) =>
-    function patchedReq(...args) {
-      // args can be (url, options, cb), (options, cb), or (url, cb)
-      let options = {};
-      const cb =
-        typeof args[args.length - 1] === 'function'
-          ? args[args.length - 1]
-          : null;
-
-      if (typeof args[0] === 'string' || args[0] instanceof URL) {
-        const parsed = typeof args[0] === 'string' ? new URL(args[0]) : args[0];
-        options =
-          typeof args[1] === 'object' && args[1] !== null
-            ? { url: parsed.href, ...args[1] }
-            : { url: parsed.href };
-      } else {
-        options = args[0] || {};
-      }
-
-      capture({
-        url:
-          options.url ||
-          `${protocol}://${options.host || options.hostname || 'localhost'}${options.path || '/'}`,
-        headers: options.headers || {},
-        method: options.method || 'GET',
-      });
-
-      // Return a no-op request that doesn't actually connect.
-      // Use a real Readable stream so libraries that call .pipe() or
-      // .setEncoding() (e.g. xmlrpc's SAX deserializer) work correctly.
-      const fakeReq = new EventEmitter();
-      fakeReq.write = () => {};
-      fakeReq.end = () => {
-        const body =
-          '<?xml version="1.0"?><methodResponse><params>' +
-          '<param><value><string>ok</string></value></param>' +
-          '</params></methodResponse>';
-        const fakeRes = new Readable({
-          read() {
-            this.push(body);
-            this.push(null);
-          },
-        });
-        fakeRes.statusCode = 200;
-        fakeRes.headers = { 'content-type': 'text/xml' };
-        if (cb) {
-          cb(fakeRes);
-        }
-        fakeReq.emit('response', fakeRes);
-      };
-      fakeReq.setTimeout = () => fakeReq;
-      fakeReq.destroy = () => {};
-      return fakeReq;
-    };
-
-  http.request = patchedRequest(origHttpRequest, 'http');
-  https.request = patchedRequest(origHttpsRequest, 'https');
-  http.get = patchedRequest(origHttpGet, 'http');
-  https.get = patchedRequest(origHttpsGet, 'https');
-
-  // Monkey-patch global fetch
-  const origFetch = globalThis.fetch;
-  globalThis.fetch = async (input, init) => {
-    const url = typeof input === 'string' ? input : input?.url || '';
-    const headers = init?.headers || input?.headers || {};
-    capture({
-      url,
-      headers:
-        headers instanceof Headers
-          ? Object.fromEntries(headers.entries())
-          : headers,
-      method: init?.method || input?.method || 'GET',
-    });
-    return new Response('{}', { status: 200, headers: {} });
-  };
-
   const bundle = {
     authData: placeholderAuthData,
     inputData: {},
@@ -766,7 +682,9 @@ const runTestFunctionSurvival = async (
   };
 
   try {
-    await withProxiedEnv(() => testFn(stubZ, bundle));
+    await withHttpCapture(capture, () =>
+      withProxiedEnv(() => testFn(stubZ, bundle)),
+    );
   } catch (err) {
     // If a request was captured before the error, use its template.
     // Many test functions crash parsing the stub response (e.g.,
@@ -776,12 +694,6 @@ const runTestFunctionSurvival = async (
       return { template: extractTemplate(capturedReq), requestMade: true };
     }
     return { template: {}, requestMade: false, error: err.message };
-  } finally {
-    http.request = origHttpRequest;
-    https.request = origHttpsRequest;
-    http.get = origHttpGet;
-    https.get = origHttpsGet;
-    globalThis.fetch = origFetch;
   }
 
   if (!capturedReq) {
